requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
ultralytics>=8.3.0
pillow>=10.0.0
numpy>=1.24.0
torch>=2.0.0
torchvision>=0.15.0
